Side menu fix
