+++ src/views/GameView.vue
@@ -3,27 +3,8 @@
 
   <HelpOverlay></HelpOverlay>
 
-  <n-layout has-sider v-if="state.token && state.connected" :class="getDirection()">
-    <n-layout-sider id="bottom-left" :collapsed-width="0" :on-update:collapsed="openCloseSider" :collapsed="state.options.hideSidebar" v-if="!state.options.hideSidebar">
-      <div class="stats-area">
-        <PlayerStats></PlayerStats>
-        <div class="bottom-area" >
-          <div class="battle-area" v-show="state.gameState.battle.active">
-            <BattleControls></BattleControls>
-          </div>
-          <div class="map-actions">
-            <MapActions v-show="!state.gameState.battle.active"></MapActions>
-          </div>
-          <div class="map-area" v-show="!state.gameState.battle.active" v-if="state.options.showMapArea">
-            <MoveControls></MoveControls>
-            <n-icon class="map-icon" v-on:click="toggleMap()"><MapOutlined /></n-icon>
-            <MiniMap></MiniMap>
-          </div>
-          <QuickStats></QuickStats>
-        </div>
-      </div>
-    </n-layout-sider>
-
+  <n-layout has-sider v-if="state.token && state.connected" :class="getDirection()" :sider-placement="state.options.swapControls ? 'right' : 'left'">
+    <SideMenu v-if="!state.options.swapControls"></SideMenu>
     <n-layout>
       <ModalCard></ModalCard>
       <MapModal></MapModal>
@@ -32,38 +13,21 @@
       </div>
       <KeyboardInput></KeyboardInput>
     </n-layout>
+    <SideMenu v-if="state.options.swapControls"></SideMenu>
   </n-layout>
 </template>
 
 <script setup>
 import { state } from '@/composables/state'
-
-import { NLayout, NLayoutSider, NIcon } from 'naive-ui'
+import { NLayout } from 'naive-ui'
 
 import LineOutput from '@/components/main-area/LineOutput.vue'
 import KeyboardInput from '@/components/main-area/KeyboardInput.vue'
 import LogoutModal from '@/components/modals/LogoutModal.vue'
-import MiniMap from '@/components/side-menu/MiniMap.vue'
-import PlayerStats from '@/components/side-menu/PlayerStats.vue'
-import BattleControls from '@/components/side-menu/BattleControls.vue'
-import MoveControls from '@/components/side-menu/MoveControls.vue'
-import QuickStats from '@/components/side-menu/QuickStats.vue'
-import MapActions from '@/components/side-menu/MapActions.vue'
 import HelpOverlay from '@/components/HelpOverlay.vue'
 import ModalCard from '@/components/modals/ModalCard'
 import MapModal from '@/components/modals/MapModal'
-import MapOutlined from '@vicons/material/MapOutlined'
-
-import { command_ids } from '@/composables/constants/command_ids'
-import { useWindowHandler } from '@/composables/window_handler'
-import { useWebSocket } from '@/composables/web_socket'
-
-const { triggerResize } = useWindowHandler()
-const { cmd } = useWebSocket()
-
-function openCloseSider () {
-  triggerResize()
-}
+import SideMenu from '@/components/side-menu/SideMenu'
 
 try {
   const options = JSON.parse(localStorage.getItem('options'))
@@ -73,13 +37,6 @@ try {
 } catch (err) {
   console.log(err)
   localStorage.setItem('options', '')
-}
-
-function toggleMap() {
-  if (!state.modals.mapModal) {
-    cmd('map', command_ids.MAP)
-  }
-  state.modals.mapModal = !state.modals.mapModal
 }
 
 function getDirection() {
@@ -126,14 +83,6 @@ function getDirection() {
     flex-grow: 1;
     min-height: 0;
   }
-}
-
-.left-side > .n-layout-scroll-container {
-  flex-flow: row nowrap !important;
-}
-
-.right-side > .n-layout-scroll-container {
-  flex-flow: row-reverse nowrap !important;
 }
 
 .n-layout {
